Finding problems with image frame tearing. WIP

+++ ract/cubes/_c_gpio_impl.py
@@ -73,6 +73,6 @@ class CffiGpioCube(object):
         :return:
         """
         for layer in self._flattened_frame_layers:
-            np.put(self.reordered_layer, self.mapping, layer.astype(int))
+            np.put(self.reordered_layer, self.mapping, layer.astype(np.uint16))
             c_pointer = ffi.cast("uint16_t *", ffi.from_buffer(self.reordered_layer))
             ract_dma.clock_in_grey_scale_data(c_pointer, len(self.reordered_layer))
